requirements.txt
# Add your project dependencies here 
RPi.GPIO>=0.7.0
numpy>=1.21.0 